backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
appscript==1.2.5
attrs==23.2.0
blinker==1.8.2
bottle==0.12.25
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==3.4.7
distro==1.9.0
et-xmlfile==1.1.0
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
macholib==1.16.3
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.28.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
proxy_tools==0.1.0
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-WebKit==10.3.1
PySocks==1.7.1
python-dotenv==1.0.1
pywebview==5.1
requests==2.32.3
selenium==4.23.0
setuptools==72.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
xlwings==0.31.10
yarl==1.9.4
